services/summarizer/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-pubsub
